+++ frontend/src/features/Signup/components/IntroPage/IntroPage.module.css
@@ -1,0 +1,38 @@
+.introTitle {
+    width: 100%;
+    height: 200px;
+    display: flex;
+    align-items: flex-end;
+    font-size: 30px;
+    color: #0046FF;
+    animation: up 2s;
+    font-family: 'LINESeedKR-Bd';
+    text-shadow: 0px 0px 5px rgba(0, 0, 0, 0.1);
+}
+
+.introText {
+    font-size: 18px;
+    color: #8A8A8A;
+    animation: up 2s;
+}
+
+.introText > span{
+    color: #000000;
+}
+
+.introFontBox {
+    width: 70px;
+    height: 20px;
+    position: relative;
+    top: -20px;
+    background-color: rgba(167, 190, 255, 0.3);
+}
+
+@keyframes up {
+    from {
+        transform: translateY(30px);
+    }
+    to {
+        transform: translateY(0px);
+    }
+}
