+++ frontend/src/features/Signup/components/PasswordPage/PasswordPage.module.css
@@ -1,0 +1,36 @@
+.inputTitle {
+    margin: 10px;
+    font-size: 18px;
+    color: #363636;
+    text-shadow: 0px 0px 5px rgba(0, 0, 0, 0.1);
+}
+
+.inputBox {
+    height: 50px;
+    margin: 10px;
+    border-bottom: 1px solid #D9D9D9;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.inputBox img {
+    width: 25px;
+    margin: 0px 10px;
+}
+
+.inputBox input {
+    width: calc(100% - 50px);
+    height: 100%;
+    padding: 0px;
+    border: none;
+    font-family: 'LINESeedKR-Rg';
+    font-size: 17px;
+    outline: none;
+}
+
+.orderText {
+    font-size: 13px;
+    margin: 0px 10px;
+    color:#0046FF;
+}
