+++ frontend/src/features/Signup/components/RolePage/RolePage.module.css
@@ -1,0 +1,63 @@
+.inputTitle {
+    margin: 10px;
+    font-size: 18px;
+    color: #363636;
+    text-shadow: 0px 0px 5px rgba(0, 0, 0, 0.1);
+}
+
+.roleBox {
+    height: 100px;
+    margin: 0px 10px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+.roleText {
+    width: 100%;
+    margin-bottom: 10px;
+    color: #8A8A8A;
+    text-shadow: 0px 0px 5px rgba(0, 0, 0, 0.1);
+}
+
+.roleButtonBox {
+    width: 100%;
+    height: 60%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 7px;
+    box-shadow: 0px 4px 10px 2px rgba(0, 0, 0, 0.1);
+}
+
+.roleButtonBox button {
+    width: 46%;
+    height: 80%;
+    border-radius: 10px;
+    border: none;
+    outline: none;
+    font-size: 18px;
+    transition: all 0.2s linear;
+    font-family: 'LINESeedKR-Bd';
+}
+
+.roleButtonBox button div {
+    line-height: 53px;
+}
+
+.roleButtonBox button:hover {
+    cursor: pointer;
+    background-color: #0036cc;
+    color: white;
+}
+
+.clicked {
+    background-color: #0046FF;
+    color: white;
+}
+
+.nonClicked {
+    color: #8A8A8A;
+    background-color: white;
+}
