+++ frontend/src/pages/SignupPage/SignupPage.module.css
@@ -18,22 +18,22 @@
 /* common */
 
 /* bold체 */
-.introTitle, .introText > span, .buttonBox button, .roleButtonBox button {
+.introText > span, .buttonBox button {
     font-family: 'LINESeedKR-Bd';
 }
 
 /* text shadow */
-.introTitle, .introText, .buttonBox button, .inputTitle, .roleText {
+.introText, .buttonBox button {
     text-shadow: 0px 0px 5px rgba(0, 0, 0, 0.1);
 }
 
 /* box shadow */
-.buttonBox button, .roleButtonBox,  .phoneBox button:hover, .certifBox button:hover {
+.buttonBox button{
     box-shadow: 0px 4px 10px 2px rgba(0, 0, 0, 0.1);
 }
 
 /* animation */
-.buttonBox button, .progressBar, .roleButtonBox button, .show, .unshow, .phoneBox button, .certifBox button {
+.buttonBox button, .progressBar{
     transition: all 0.2s linear;
  }
 
@@ -85,229 +85,6 @@
     color: #8A8A8A;
 }
 
-@keyframes up {
-    from {
-        transform: translateY(30px);
-    }
-    to {
-        transform: translateY(0px);
-    }
-}
-
-/* intro 페이지 */
-.introTitle {
-    width: 100%;
-    height: 200px;
-    display: flex;
-    align-items: flex-end;
-    font-size: 30px;
-    color: #0046FF;
-    animation: up 2s;
-}
-
-
-.introText {
-    font-size: 18px;
-    color: #8A8A8A;
-    animation: up 2s;
-}
-
-.introText > span{
-    color: #000000;
-}
-
-.introFontBox {
-    width: 70px;
-    height: 20px;
-    position: relative;
-    top: -20px;
-    background-color: rgba(167, 190, 255, 0.3);
-}
-
-/* 이름 작성 페이지 */
-.inputTitle {
-    margin: 10px;
-    font-size: 18px;
-    color: #363636;
-}
-
-.inputBox {
-    height: 50px;
-    margin: 10px;
-    border-bottom: 1px solid #D9D9D9;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.inputBox img {
-    width: 25px;
-    margin: 0px 10px;
-}
-
-.inputBox input {
-    width: calc(100% - 50px);
-    height: 100%;
-    padding: 0px;
-    border: none;
-    font-family: 'LINESeedKR-Rg';
-    font-size: 17px;
-    outline: none;
-}
-
-/* 전화번호 페이지 */
-.phoneBox {
-    height: 50px;
-    margin: 10px;
-    border-bottom: 1px solid #D9D9D9;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.phoneBox img {
-    width: 25px;
-    margin: 0px 10px;
-}
-
-.phoneBox input {
-    width: calc(100% - 150px);
-    height: 100%;
-    padding: 0px;
-    border: none;
-    font-family: 'LINESeedKR-Rg';
-    font-size: 17px;
-    outline: none;
-}
-
-.phoneBox button {
-    width: 90px;
-    height: 40px;
-    margin-left: 10px;
-    outline: none;
-    border: none;
-    font-family: 'LINESeedKR-Rg';
-    font-size: 13px;
-    color:white;
-    border-radius: 5px;
-    line-height: 40px;
-}
-
-.phoneBox button:hover {
-    cursor: pointer;
-}
-
-.show {
-    opacity: 1;
-}
-
-.unshow {
-    opacity: 0;
-}
-
-.orderText {
-    font-size: 13px;
-    margin: 0px 10px;
-    color:#0046FF;
-}
-
-.certifBox {
-    height: 50px;
-    margin: 10px;
-    border-bottom: 1px solid #D9D9D9;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-
-.certifBox input {
-    width: calc(100% - 160px);
-    height: 100%;
-    padding: 0px;
-    border: none;
-    font-family: 'LINESeedKR-Rg';
-    font-size: 17px;
-    outline: none;
-}
-
-.certifBox button {
-    width: 90px;
-    height: 40px;
-    margin-left: 10px;
-    outline: none;
-    border: none;
-    font-family: 'LINESeedKR-Rg';
-    font-size: 15px;
-    color:white;
-    background-color: #0046FF;
-    border-radius: 5px;
-    line-height: 40px;
-}
-
-.certifBox div {
-    width: 50px;
-    margin-left: 5px;
-}
-
-.certifBox button:hover {
-    cursor: pointer;
-}
-
-/* 역할 페이지 */
-.roleBox {
-    height: 100px;
-    margin: 0px 10px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.roleText {
-    width: 100%;
-    margin-bottom: 10px;
-    color: #8A8A8A;
-}
-
-.roleButtonBox {
-    width: 100%;
-    height: 60%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 7px;
-}
-
-.roleButtonBox button {
-    width: 46%;
-    height: 80%;
-    border-radius: 10px;
-    border: none;
-    outline: none;
-    font-size: 18px;
-}
-
-.roleButtonBox button div {
-    line-height: 53px;
-}
-
-.roleButtonBox button:hover {
-    cursor: pointer;
-    background-color: #0036cc;
-    color: white;
-}
-
-.clicked {
-    background-color: #0046FF;
-    color: white;
-}
-
-.nonClicked {
-    color: #8A8A8A;
-    background-color: white;
-}
-
 /* 하단 버튼 박스 */
 .buttonBox div {
     text-align: center;
@@ -334,11 +111,5 @@
     border-radius: 10px;
     font-size: 20px;
     color: white;
-    /* background-color: #0046FF; */
     background-color: #D9D9D9;
 }
-
-/* .buttonBox button:hover {
-    box-shadow: 0px 4px 10px 6px rgba(0, 0, 0, 0.3);
-    cursor: pointer;
-} */
